+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [PR #2021](https://github.com/rqlite/rqlite/pull/2021): Fix typos in Help strings. Thanks @alexandear.
 - [PR #2024](https://github.com/rqlite/rqlite/pull/2024): Improve variable name in SQL parsing code.
 - [PR #2025](https://github.com/rqlite/rqlite/pull/2025): Optimize need-to-parse logic.
+- [PR #2026](https://github.com/rqlite/rqlite/pull/2026): Recover from any panic in SQL parser.
 
 ## v8.36.5 (January 11th 2025)
 ### Implementation changes and bug fixes